+++ code/lib/angular-compodoc/build-config.ts
@@ -1,0 +1,18 @@
+import type { BuildEntries } from '../../../scripts/build/utils/entry-utils.ts';
+
+const config: BuildEntries = {
+  entries: {
+    browser: [
+      {
+        exportEntries: ['.'],
+        entryPoint: './src/index.ts',
+      },
+      {
+        exportEntries: ['./browser'],
+        entryPoint: './src/browser.ts',
+      },
+    ],
+  },
+};
+
+export default config;
