+++ code/lib/angular-compodoc/project.json
@@ -1,0 +1,10 @@
+{
+  "name": "angular-compodoc",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "targets": {
+    "compile": {},
+    "check": {}
+  },
+  "tags": ["library"]
+}
