+++ code/lib/angular-compodoc/src/typings.d.ts
@@ -1,0 +1,4 @@
+/** Where `setCompodocJson` parks Compodoc's `documentation.json` for the preview to read back. */
+declare var __STORYBOOK_COMPODOC_JSON__: any;
+
+declare var FEATURES: import('storybook/internal/types').StorybookConfigRaw['features'];
